fix(auth): 修复强制修改密码后未提示成功的问题 (Issue #2230)
问题：修改密码成功后，页面未显示成功提示，需点击 retry 才进入系统。

原因：
1. React Query 缓存未失效，导致 403 失败的查询没有自动重试
2. Vite 构建时 hooks 和 components chunk 之间存在循环依赖，
   导致 useToastStore 在运行时存在多个实例

修复：
- 新建 frontend/src/store/toastStore.ts 打破循环依赖
- 修改 Toast.tsx 从 store/toastStore 导入 useToastStore
- 在 ForceChangePasswordModal.tsx 中添加 toast.success 调用
- 在 useAuth.ts 的 onSuccess 中添加 invalidateQueries 调用
- 添加 passwordChangedSuccess 多语言键值（英/中/日/韩）

原始提交: 52b2fcfda6bcf8128e326398bdec2d5c7ec30fea
来源: papercatno1/open-eduace fork
