+++ frontend/src/components/common/Toast.tsx
@@ -13,18 +13,12 @@
 
 import React, { useEffect, useState, useCallback } from 'react';
 import { createPortal } from 'react-dom';
-import { create } from 'zustand';
 import { cn } from '@/utils';
+import { useToastStore } from '@/store/toastStore';
+import type { ToastType, ToastData, ToastStore } from '@/store/toastStore';
 
-export type ToastType = 'success' | 'error' | 'warning' | 'info';
-
-export interface ToastData {
-  id: string;
-  type: ToastType;
-  title: string;
-  message?: string;
-  duration?: number;
-}
+// Re-export types for backward compatibility with existing imports.
+export type { ToastType, ToastData };
 
 interface ToastProps extends ToastData {
   onClose: (id: string) => void;
@@ -121,37 +115,9 @@ export const ToastContainer: React.FC<ToastContainerProps> = ({
 };
 
 /**
- * Global toast store - single source of truth for all toasts in the app.
+ * Global toast store - imported from store/toastStore.ts to avoid circular
+ * chunk dependencies between hooks and components.
  */
-interface ToastStore {
-  toasts: ToastData[];
-  addToast: (toast: Omit<ToastData, 'id'>) => string;
-  removeToast: (id: string) => void;
-  clearToasts: () => void;
-  success: (title: string, message?: string, duration?: number) => string;
-  error: (title: string, message?: string, duration?: number) => string;
-  warning: (title: string, message?: string, duration?: number) => string;
-  info: (title: string, message?: string, duration?: number) => string;
-}
-
-const generateId = (): string => `toast-${Date.now()}-${Math.random().toString(36).substr(2, 9)}`;
-
-export const useToastStore = create<ToastStore>((set, get) => ({
-  toasts: [],
-  addToast: (toast) => {
-    const id = generateId();
-    set((state) => ({ toasts: [...state.toasts, { ...toast, id }] }));
-    return id;
-  },
-  removeToast: (id) => set((state) => ({ toasts: state.toasts.filter((t) => t.id !== id) })),
-  clearToasts: () => set({ toasts: [] }),
-  success: (title, message, duration) =>
-    get().addToast({ type: 'success', title, message, duration }),
-  error: (title, message, duration) => get().addToast({ type: 'error', title, message, duration }),
-  warning: (title, message, duration) =>
-    get().addToast({ type: 'warning', title, message, duration }),
-  info: (title, message, duration) => get().addToast({ type: 'info', title, message, duration }),
-}));
 
 /**
  * Stable singleton API. Returning the same object reference every call keeps
