+++ frontend/src/i18n/index.ts
@@ -216,6 +216,7 @@ export const translations: Record<Language, Translations> = {
     enterCurrentPassword: 'Enter current password',
     enterNewPassword: 'Enter new password',
     failedToChangePassword: 'Failed to change password',
+    passwordChangedSuccess: 'Password changed successfully',
     failedToSaveUser: 'Failed to save user',
     enterUsername: 'Enter username',
     enterEmail: 'Enter email',
@@ -1971,6 +1972,7 @@ export const translations: Record<Language, Translations> = {
     enterCurrentPassword: '请输入当前密码',
     enterNewPassword: '请输入新密码',
     failedToChangePassword: '修改密码失败',
+    passwordChangedSuccess: '密码修改成功',
     failedToSaveUser: '保存用户失败',
     enterUsername: '请输入用户名',
     enterEmail: '请输入邮箱',
@@ -3787,6 +3789,7 @@ export const translations: Record<Language, Translations> = {
     enterCurrentPassword: '現在のパスワードを入力',
     enterNewPassword: '新しいパスワードを入力',
     failedToChangePassword: 'パスワードの変更に失敗しました',
+    passwordChangedSuccess: 'パスワードが正常に変更されました',
     failedToSaveUser: 'ユーザーの保存に失敗しました',
     enterUsername: 'ユーザー名を入力',
     enterEmail: 'メールを入力',
@@ -5372,6 +5375,7 @@ export const translations: Record<Language, Translations> = {
     enterCurrentPassword: '현재 비밀번호 입력',
     enterNewPassword: '새 비밀번호 입력',
     failedToChangePassword: '비밀번호 변경 실패',
+    passwordChangedSuccess: '비밀번호가 성공적으로 변경되었습니다',
     failedToSaveUser: '사용자 저장 실패',
     enterUsername: '사용자 이름 입력',
     enterEmail: '이메일 입력',
